// frontend/src/lib/api/models/collectionDto.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Snippet.Web | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CollectionDto {
	id: string;
	userId: string;
	name: string;
	/** @nullable */
	description: string | null;
	/** @nullable */
	color: string | null;
	/** @nullable */
	icon: string | null;
	sortOrder: number;
	snippetCount: number;
	createdAt: string;
}